+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Parsing of repeated label definitions.
 - Consolidation of `in` as identifier in `class operator In` construct.
 - Duplicated line attribution of mid-line compiler directives.
+- Parsing of `class helper`s with parent a type.
 
 ## [0.3.0] - 2024-05-29